Mapped Strings to Components
Turned out not to be as helpful as I thought.  I can’t write a
generalized write method for tables (since they need input as to which
cell to get), and a large chunk of the code comes from reading in table
values.

Will talk to Carlos about this, but will probably end up reverting this
commit, since I think it just obfuscates the code rather than clearing
it up.